Extract noop analytics implementation
